Add files via upload
You are given an array A1,A2,…,AN of length N sorted in non-decreasing order. Your task is to remove all the duplicates and find the sorted increasing array of distinct elements consisting of all distinct elements present in A.

Input Format
The first line of input will contain a single integer T, denoting the number of test cases.
The first line of each test case contains an integer N - the length of the array A.
The second line of each test case contains N space-separated integers A1,A2,…,AN .
Output Format
For each test case, output two lines:

The first line should contain a single integer M - the size of the array.
The second line should contain M space-separated integers denoting the elements of the array.
Constraints
1≤T≤100
1≤N≤10^5
1≤Ai≤10^9
 
The sum of N over all test cases won't exceed 2⋅10^5.
Sample 1:
Input
3
2
5 10
4
1 5 5 10
5
4 4 6 6 8
Output
2
5 10 
3
1 5 10 
3
4 6 8 
Explanation:
Test case 1: Distinct elements will be 5, 10. So the array will be [5,10].

Test case 2: Distinct elements will be 1, 5, 10. So the array will be [1,5,10].

Test case 3: Distinct elements will be 4, 6, 8. So the array will be [4,6,8].