Add files via upload
You are given an array of integers. Your task is to sort the array such that all odd numbers come before all even numbers while maintaining the original relative order of the odd and even numbers.

Input Format
The first line contains an integer N (the array size).
The second line contains N space-separated integers representing the elements of the array.
Output Format
Output the sorted array where all odd numbers appear before all even numbers, maintaining the relative order.

Constraints
1≤N≤10^5
0≤a[i]≤10^9
 
Sample 1:
Input
5
3 1 2 4 5
Output
3 1 5 2 4
Explanation:
In the given array, the odd numbers are 3,1 and 5, and the even numbers are 2 and 4. After sorting by parity while maintaining the original order, the result is 3,1,5,2,4.

Sample 2:
Input
6
2 4 6 1 3 5
Output
1 3 5 2 4 6
Explanation:
In the given array, the odd numbers are 1,3 and 5, and the even numbers are 2,4 and 6. After sorting by parity while maintaining the original order, the result is 1,3,5,2,4,6.