Add files via upload
You are given two arrays of positive integers, arr1 and arr2, of sizes n and m respectively. Your task is to find the largest common element that appears in both arrays. If no common element exists, return −1.

Input Format
The first line contains two integers n and m, the sizes of arr1 and arr2.
The second line contains n space-separated integers representing the elements of arr1.
The third line contains m space-separated integers representing the elements of arr2.
Output Format
Output a single integer which is the largest common element in both arrays or −1 if no such element exists.

Constraints
1≤n,m≤10^5
0≤arr1[i],arr2[i]≤10^9
 
Sample 1:
Input
5 6
1 3 4 6 7
2 3 5 6 7 8
Output
7
Explanation:
In the given arrays, the common elements are 3, 6, and 7. The largest common element is 7.

Sample 2:
Input
Output
4 3
1 2 3 4
5 6 7
-1
Explanation:
In the given arrays, there are no common elements. Therefore, the output is -1.