Add files via upload
Let's suppose we are given an integer N and we have to calculate the sum of first N natural numbers. This problem can easily be broken down to smaller parts of the same nature.

Assuming Sum(N) denotes the sum of first N natural numbers, we can saySum(5)=5+Sum(4),
which can be broken down further to saySum(5)=5+4+Sum(3) and so on.

These broken down versions are generally called sub-problems. We break the task into sub-problems until the task becomes so small that the solution becomes trivial. Like in the above problem if we keep breaking this into smaller problems, it reaches 1, whereSum(1)=1.

Thus the solution becomes trivial.
From the above example, we can say that:Sum(N)=N+Sum(N−1), until N>1